sys: kern: arch: rv32: break trap routines into their own module
